# contracts/openapi/emcore-identity-access-api/v1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EMCORE Identity \u0026 Access API",
    "description": "Manages user registration, credential authentication, multi-factor authentication (MFA), step-up authorization workflows, JWT session token issuance, session revocation, workload service client identities, JWKS verification keys, and administrative user security status locking. Owns authentication and cryptographic tokens; does not own tenant role definitions or business permissions directly.\n\n**Service Ownership:** Identity \u0026 Access Security Team\n**Intended Consumers:** Platform clients, mobile apps, gateway middleware, and federated service consumers\n**Operational Mode:** Primarily synchronous HTTP API; critical domain mutations emit versioned outbox events and may operate under eventual consistency.\n**Confidentiality:** Public/Enterprise Consumer Contract.",
    "termsOfService": "https://emcore.platform/terms-of-service",
    "contact": {
      "name": "EMCORE Architecture \u0026 Support Team",
      "url": "https://emcore.platform/support",
      "email": "support@emcore.platform"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "EMCORE Development Standalone Host Server"
    },
    {
      "url": "https://api.emcore.platform/v1",
      "description": "EMCORE Platform Target Server"
    }
  ],
  "paths": {
...
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "ServiceClient"
        ],
        "operationId": "postApiV1AuthToken",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTokenRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/identity/service-clients/register": {
      "post": {
        "tags": [
          "ServiceClient"
        ],
        "operationId": "postApiV1IdentityServiceClientsRegister",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterServiceClientRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterServiceClientRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RegisterServiceClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterServiceClientResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/identity/service-clients/{id}/rotate": {
      "post": {
        "tags": [
          "ServiceClient"
        ],
        "operationId": "postApiV1IdentityServiceClientsIdRotate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateServiceClientCredentialRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateServiceClientCredentialRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RotateServiceClientCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotateServiceClientCredentialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/identity/service-clients/credentials/revoke": {
      "post": {
        "tags": [
          "ServiceClient"
        ],
        "operationId": "postApiV1IdentityServiceClientsCredentialsRevoke",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeServiceClientCredentialRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeServiceClientCredentialRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RevokeServiceClientCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/identity/service-clients/{id}/credentials": {
      "get": {
        "tags": [
          "ServiceClient"
        ],
        "operationId": "getApiV1IdentityServiceClientsIdCredentials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceClientCredentialDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "code": "AUTH_REQUIRED",
                  "detail": "Valid authentication token or credentials were not provided.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "EmcoreProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI reference identifying the problem type.",
            "example": "https://emcore.platform/errors/400"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary of the error.",
            "example": "Validation failed"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "format": "int32",
            "example": 400
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code.",
            "example": "VALIDATION_ERROR"
          },
          "detail": {
            "type": "string",
            "description": "Detailed diagnostic explanation of the error condition.",
            "example": "One or more request parameters failed structural validation."
          },
          "traceId": {
            "type": "string",
            "description": "Distributed trace identifier.",
            "example": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
          },
          "requestId": {
            "type": "string",
            "description": "Enterprise request tracking identifier.",
            "example": "req_01HPX7K7R5YZ2X90WY"
          },
          "correlationId": {
            "type": "string",
            "description": "Distributed correlation identifier.",
            "example": "cor_01HPX7K7R5YZ2X90WY"
          },
          "errors": {
            "type": "object",
            "description": "Property-level validation failure details."
          }
        },
        "description": "RFC 7807 standardized problem details error payload."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "description": "Current operational or workflow lifecycle state indicator.",
            "format": "int32",
            "nullable": true,
            "example": "Active"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "description": "RFC 7807 standardized problem details error payload."
      },
...
      "RegisterServiceClientRequest": {
        "required": [
          "clientId",
          "scopes"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiryDays": {
            "type": "integer",
            "format": "int32",
            "default": 365
          }
        }
      },
      "RegisterServiceClientResponse": {
        "required": [
          "id",
          "clientId",
          "clientSecret",
          "keyId",
          "expiresAtUtc",
          "scopes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Opaque unique identity reference identifier (ULID/UUID format). Must not be treated as a sequential database key.",
            "example": "usr_01HPX7K7R5YZ2X90WY0001"
          },
          "clientId": {
            "type": "string"
          },
          "clientSecret": {
            "type": "string"
          },
          "keyId": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "RevokeServiceClientCredentialRequest": {
        "required": [
          "credentialId"
        ],
        "type": "object",
        "properties": {
          "credentialId": {
            "type": "string"
          }
        }
      },
      "RotateServiceClientCredentialRequest": {
        "required": [
          "serviceClientId"
        ],
        "type": "object",
        "properties": {
          "serviceClientId": {
            "type": "string"
          },
          "expiryDays": {
            "type": "integer",
            "format": "int32",
            "default": 365
          }
        },
        "nullable": true
      },
...
      "ServiceTokenRequest": {
        "required": [
          "clientId",
          "clientSecret"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "clientSecret": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "default": null,
            "nullable": true
          }
        }
      },
      "ServiceTokenResponse": {
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Cryptographically secure verification or session authorization token. Treated as sensitive credential.",
            "example": "sample_sanitized_token_value"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "tokenType": {
            "type": "string",
            "default": "Bearer"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PRODUCTION GATEWAY JWT VALIDATION: NOT IMPLEMENTED. PRODUCTION FALLBACK TO TEST AUTH: PROHIBITED. DEVELOPMENT ONLY token verification. DEFERRED.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ClientCredentials": {
        "type": "oauth2",
        "description": "OAuth-style client credentials for service identities. Requires Client ID and Secret to obtain a short-lived bearer token from POST /api/v1/auth/token.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://auth.emcore.platform/api/v1/auth/token",
            "scopes": {
              "api.access": "General API service access",
              "admin.write": "Elevated administrative state modifications"
            }
          }
        }
      },
      "StepUpToken": {
        "type": "apiKey",
        "description": "Identity Step-Up Flow: IMPLEMENTED. Downstream Step-Up Enforcement: NOT VERIFIED / NOT IMPLEMENTED.",
        "name": "X-StepUp-Token",
        "in": "header"
      },
      "WebhookHmac": {
        "type": "apiKey",
        "description": "PLANNED \u2014 NOT IMPLEMENTED.",
        "name": "X-Signature-256",
        "in": "header"
      }
    }
  },
...
}